Fisher utilization distributions

A RasterStack of utilization distributions for a single Fisher in New York state over the course of nine nights in April 2011. Cell values represent the probability density that the individual was found at a given location between sunset and sunrise and the nine layers represent nine nights of data. Utilization distributions were generated using the adehabitatHR package from GPS tracking data.

fisher_ud

Format

An object of class RasterStack of dimension 176 x 177 x 9.

Source

https://www.datarepository.movebank.org/handle/10255/move.330

References

LaPoint, S., P. Gallery, M. Wikelski, R. Kays. 2013. Animal behavior, cost-based corridor models, and real corridors. Landscape Ecology 28(8): 1615–1630. https://doi.org/10.1007/s10980-013-9910-0

LaPoint, S., P. Gallery, M. Wikelski, R. Kays. 2013. Data from: Animal behavior, cost-based corridor models, and real corridors. Movebank Data Repository. https://doi.org/10.5441/001/1.2tp2j43g