Transform a raster stack to intensity values

Transform a raster stack to intensity values

metrics_pull(x)

Arguments

x
RasterStack with each layer typically corresponding to the distribution of an individual or species at a given time interval.

Value

A RasterStack resulting from converting the input stack to intensity values by dividing each layer by the maximum pixel value in that layer. These layer maxima are stored as the "maximum" attribute.

See also

Examples

# load elephant data
data("elephant_ud")
r <- metrics_pull(elephant_ud)
print(r)

#> class      : RasterBrick 
#> dimensions : 218, 220, 47960, 4  (nrow, ncol, ncell, nlayers)
#> resolution : 500, 500  (x, y)
#> extent     : 1472441, 1582441, -2379329, -2270329  (xmin, xmax, ymin, ymax)
#> crs        : +proj=moll +lon_0=0 +x_0=0 +y_0=0 +ellps=WGS84 +units=m +no_defs 
#> source     : memory
#> names      : year1, year2, year3, year4 
#> min values :     0,     0,     0,     0 
#> max values :     1,     1,     1,     1 
#> 

# maximum values for each layer stored as an attribute
attr(r, "maximum")

#> [1] 0.003741765 0.004792186 0.004750540 0.003533928