eBird Status & Trends Field Sparrow occurrence probability

A RasterStack of the expected probability of occurrence of Field Sparrow from the eBird Status & Trends project. Each of the 12 layers in the stack represent the estimated occurrence for a given month of the year over a regular grid of points covering the full range of the species. To reduce file size, these data have been aggregated from the native 2.96 km spatial resolution and weekly temporal resolution to monthly, 14.8 km resolution.

fiespa_occ

Format

An object of class RasterStack of dimension 193 x 225 x 12.

Source

https://ebird.org/science/status-and-trends

Details

For further details on these data, and to access the data for more species, consult the documentation for the ebirdst package.

References

Fink, D., T. Auer, A. Johnston, M. Strimas-Mackey, O. Robinson, S. Ligocki, B. Petersen, C. Wood, I. Davies, B. Sullivan, M. Iliff, S. Kelling. 2020. eBird Status and Trends, Data Version: 2018; Released: 2020. Cornell Lab of Ornithology, Ithaca, New York. https://doi.org/10.2173/ebirdst.2018